requirements.txt
# Core HTTP
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Numerics
numpy>=1.26.0

# GUI / Charts
matplotlib>=3.8.0

# Security / credential encryption
cryptography>=42.0.0

# Backtesting performance metrics
scipy>=1.12.0

# Terminal UI
rich>=13.0.0

# PDF manual generation (generate_manual.py)
reportlab>=3.6.0
